requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-spectacular
drf-spectacular[sidecar]
